More hash tests
