+++ tracer/test/Datadog.Trace.Instrumented.Iast.Unit.Tests/Vulnerabilities/WeakHashing/HMACMD5AspectTests.cs
@@ -1,0 +1,53 @@
+// <copyright file="HMACMD5AspectTests.cs" company="Datadog">
+// Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
+// This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
+// </copyright>
+
+using System;
+using System.IO;
+using System.Security.Cryptography;
+using Moq;
+using Xunit;
+
+namespace Datadog.Trace.Instrumented.Iast.Unit.Tests.Vulnerabilities.WeakHashing;
+public class HMACMD5AspectTests : InstrumentationTestsBase
+{
+    [Fact]
+    public void GivenAHMACMD5_WhenComputeHash_VulnerabilityIsLogged()
+    {
+        new HMACMD5().ComputeHash(new Mock<Stream>().Object);
+        AssertVulnerable();
+    }
+
+    [Fact]
+    public void GivenAHMACMD5_WhenComputeHash_VulnerabilityIsLogged2()
+    {
+        new HMACMD5().ComputeHash(new byte[] { });
+        AssertVulnerable();
+    }
+
+    [Fact]
+    public void GivenAHMACMD5_WhenComputeHash_ExceptionIsThrown()
+    {
+        Assert.Throws<ArgumentException>(() => new HMACMD5().ComputeHash(new byte[] { }, 10, 10));
+    }
+
+    [Fact]
+    public void GivenAHMACMD5_WhenComputeHashNullStream_ExceptionIsThrown2()
+    {
+        Assert.Throws<NullReferenceException>(() => new HMACMD5().ComputeHash(((Stream)null)));
+    }
+
+    [Fact]
+    public void GivenAHMACMD5_WhenComputeHash_VulnerabilityIsLogged3()
+    {
+        new HMACMD5().ComputeHash(new byte[] { 5, 5, 5 }, 0, 2);
+        AssertVulnerable();
+    }
+
+    [Fact]
+    public void GivenANewHMACMD5_WhenComputeHash_ExceptionIsThrown()
+    {
+        Assert.Throws<ArgumentException>(() => new HMACMD5().ComputeHash(new byte[] { 5, 5, 5 }, 10, 2));
+    }
+}
